- make styling more consistent on settings view
